+++ src/plugins/workspace/server/saved_objects/workspace_id_consumer_wrapper.ts
@@ -12,6 +12,8 @@ import {
   SavedObjectsCheckConflictsObject,
   OpenSearchDashboardsRequest,
   SavedObjectsFindOptions,
+  PUBLIC_WORKSPACE_ID,
+  WORKSPACE_TYPE,
 } from '../../../../core/server';
 
 type WorkspaceOptions = Pick<SavedObjectsBaseOptions, 'workspaces'> | undefined;
@@ -37,6 +39,15 @@ export class WorkspaceIdConsumerWrapper {
       ...(finalWorkspaces.length ? { workspaces: finalWorkspaces } : {}),
     };
   }
+
+  private isWorkspaceType(type: SavedObjectsFindOptions['type']): boolean {
+    if (Array.isArray(type)) {
+      return type.every((item) => item === WORKSPACE_TYPE);
+    }
+
+    return type === WORKSPACE_TYPE;
+  }
+
   public wrapperFactory: SavedObjectsClientWrapperFactory = (wrapperOptions) => {
     return {
       ...wrapperOptions.client,
@@ -63,8 +74,31 @@ export class WorkspaceIdConsumerWrapper {
           this.formatWorkspaceIdParams(wrapperOptions.request, options)
         ),
       delete: wrapperOptions.client.delete,
-      find: (options: SavedObjectsFindOptions) =>
-        wrapperOptions.client.find(this.formatWorkspaceIdParams(wrapperOptions.request, options)),
+      find: (options: SavedObjectsFindOptions) => {
+        const findOptions = this.formatWorkspaceIdParams(wrapperOptions.request, options);
+        if (this.isWorkspaceType(findOptions.type)) {
+          return wrapperOptions.client.find(findOptions);
+        }
+
+        // if workspace is enabled, we always find by workspace
+        if (!findOptions.workspaces || findOptions.workspaces.length === 0) {
+          findOptions.workspaces = [PUBLIC_WORKSPACE_ID];
+        }
+
+        // `PUBLIC_WORKSPACE_ID` includes both saved objects without any workspace and with `PUBLIC_WORKSPACE_ID` workspace
+        const index = findOptions.workspaces
+          ? findOptions.workspaces.indexOf(PUBLIC_WORKSPACE_ID)
+          : -1;
+        if (!findOptions.workspacesSearchOperator && findOptions.workspaces && index !== -1) {
+          findOptions.workspacesSearchOperator = 'OR';
+          // remove this deletion logic when public workspace becomes to real
+          if (this.isPermissionControlEnabled) {
+            // remove public workspace to make sure we can pass permission control validation, more details in `WorkspaceSavedObjectsClientWrapper`
+            findOptions.workspaces.splice(index, 1);
+          }
+        }
+        return wrapperOptions.client.find(findOptions);
+      },
       bulkGet: wrapperOptions.client.bulkGet,
       get: wrapperOptions.client.get,
       update: wrapperOptions.client.update,
@@ -75,5 +109,5 @@ export class WorkspaceIdConsumerWrapper {
     };
   };
 
-  constructor() {}
+  constructor(private isPermissionControlEnabled?: boolean) {}
 }